Adds initial code for a new SARS hub (#168)
* Adds initial code for a new SARS hub fixes #167

* Automatically make soft links for data directories within trackhub dir and switch bigDataUrl to relative paths

+++ dnase/trackhub/update_tracks.bash
@@ -5,15 +5,16 @@
 # This script can be called from anywhere.
 #
 # update_tracks.bash hub_type hub_target short_label long_label URLbase
-# hub_type:  Either CEGS or MAURANOLAB
+# hub_type:  One of: CEGS, MAURANOLAB, or SARS
 # hub_target:  The full path to the new hub location, which must already exist and be empty.
 # short_label, long_label:  The short and long labels that appear on the UCSC browser Track Data Hubs page. If you have multiple hubs you should make sure these are unique, so you know what you are clicking on.
-# URLbase:      The complete url to the new hub location; assumes that mapped, aggregations, and publicdata are present as subdirectories
+# URLbase:      Complete URL (e.g. https://<login:pwd>@cascade.isg.med.nyu.edu/cegs) to the directory which contains the hub directory (e.g. "trackhub", trackhub_dev", "trackhub_sars"), used for hubCheck
 #
 #
-# Standard execution of this script (for CEGS dev and MAURANOLAB):
+# Standard execution of this script:
 #/vol/cegs/src/trackhub/src_dev/update_tracks.bash CEGS /vol/cegs/public_html/trackhub_dev "CEGS Dev" "CEGS Development Hub" https://cascade.isg.med.nyu.edu/cegs
 #/vol/cegs/src/trackhub/src_prod/update_tracks.bash MAURANOLAB /home/cadlej01/public_html/trackhub "Maurano Lab" "Maurano Lab Hub" https://cascade.isg.med.nyu.edu/~cadlej01
+#/vol/mauranolab/mapped/src/dnase/trackhub/update_tracks.bash SARS /vol/mauranolab/sars/public_html/trackhub "SARS-CoV2" "SARS-CoV2 Hub" https://cascade.isg.med.nyu.edu/sars
 #
 # There are two sets of code, defined as development and production.
 # They are located here:
@@ -69,7 +70,6 @@ short_label=$3
 long_label=$4
 URLbase=$5
 
-
 # Check the inputs:
 if [[ "${hub_type}" == "CEGS" ]]; then
     customGenomeAssembly="cegsvectors"
@@ -77,8 +77,11 @@ if [[ "${hub_type}" == "CEGS" ]]; then
 elif [[ "${hub_type}" == "MAURANOLAB" ]]; then
     customGenomeAssembly="mauranolab"
     assemblyBaseDir="/vol/mauranolab"
+elif [[ "${hub_type}" == "SARS" ]]; then
+    customGenomeAssembly="NA"
+    assemblyBaseDir="/vol/mauranolab/sars"
 else
-    echo "ERROR You need to enter a valid hub type. Either: CEGS or MAURANOLAB. Exiting..."
+    echo "ERROR You need to enter a valid hub type. Either: CEGS or MAURANOLAB or SARS. Exiting..."
     exit 1
 fi
 
@@ -102,6 +105,13 @@ fi
 
 echo "Building new ${hub_type} track hub in ${hub_target}"
 
+# Make soft links in hub directory, so we can have multiple hubs with their own data sources in public_html.
+for subdir in mapped aggregations publicdata; do
+    if [[ -d "${assemblyBaseDir}/${subdir}" ]]; then
+        ln -s ${assemblyBaseDir}/${subdir} ${hub_target}/
+    fi
+done
+
 ############################################################################
 # We need a tmp directory to store intermediate files.
 
@@ -123,7 +133,7 @@ path_to_main_driver_script=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 cd ${path_to_main_driver_script}
 
 # What genomes will we be working with?
-# These are defined in the CEGS_genomes and MAURANOLAB_genomes files.
+# These are defined in the CEGS_genomes, MAURANOLAB_genomes, and SARS_genomes files.
 # Read the applicable one here:
 readarray -t genome_array < "${path_to_main_driver_script}/${hub_type}_genomes"
 echo " "
@@ -140,8 +150,9 @@ echo "Starting makeAssemblyTracks.bash"
 ./makeAssemblyTracks.bash ${path_to_main_driver_script} ${hub_target} ${TMPDIR} ${hub_type} ${customGenomeAssembly} ${assemblyBaseDir} "${genome_array[@]}"
 
 # Now construct the "flowcell" and "aggregation" tracks in TMPDIR.
+hub_dir=`basename ${hub_target}`
 echo "Starting make_tracks.bash"
-./make_tracks.bash ${TMPDIR} ${hub_type} ${path_to_main_driver_script} ${assemblyBaseDir} ${URLbase} "${genome_array[@]}"
+./make_tracks.bash ${TMPDIR} ${hub_type} ${path_to_main_driver_script} ${assemblyBaseDir} "${genome_array[@]}"
 
 ######################################################################################
 # Now copy the track information to the hub location.
@@ -176,19 +187,21 @@ update_genome () {
        cat "${TMPDIR}/MakeTrackhub_${genome}_consolidated_agg.out" >> trackDb_001.txt
     fi
     
-    # Process the "publicdata" tracks.
-    num_line=`(wc -l < "${TMPDIR}/samplesforTrackhub_${genome}_consolidated_pub.tsv")`
-    if [ ${num_line} -gt 1 ]; then
-       # If num_line == 1, then there are no publicdata tracks for this genome.
-       cat "${TMPDIR}/MakeTrackhub_${genome}_consolidated_pub.out" >> trackDb_001.txt
-    fi
-    
-    if [ "${hub_type}" = "CEGS" ]; then
-        # Process the "CEGS_byLocus" tracks.
-        num_line=`(wc -l < "${TMPDIR}/samplesforTrackhub_${genome}_consolidated_locus.tsv")`
+    if [[ "${hub_type}" != "SARS" ]]; then
+        # Process the "publicdata" tracks.
+        num_line=`(wc -l < "${TMPDIR}/samplesforTrackhub_${genome}_consolidated_pub.tsv")`
         if [ ${num_line} -gt 1 ]; then
-           # If num_line == 1, then there are no CEGS_byLocus tracks for this genome.
-           cat "${TMPDIR}/MakeTrackhub_${genome}_consolidated_locus.out" >> trackDb_001.txt
+           # If num_line == 1, then there are no publicdata tracks for this genome.
+           cat "${TMPDIR}/MakeTrackhub_${genome}_consolidated_pub.out" >> trackDb_001.txt
+        fi
+        
+        if [ "${hub_type}" = "CEGS" ]; then
+            # Process the "CEGS_byLocus" tracks.
+            num_line=`(wc -l < "${TMPDIR}/samplesforTrackhub_${genome}_consolidated_locus.tsv")`
+            if [ ${num_line} -gt 1 ]; then
+               # If num_line == 1, then there are no CEGS_byLocus tracks for this genome.
+               cat "${TMPDIR}/MakeTrackhub_${genome}_consolidated_locus.out" >> trackDb_001.txt
+            fi
         fi
     fi
 }
@@ -198,7 +211,9 @@ for i in "${genome_array[@]}"; do
 done
 
 # Move the GC percentage file:
-cp ${TMPDIR}/assembly_tracks/${customGenomeAssembly}.gc.bw ${hub_target}/${customGenomeAssembly}/data
+if [ -f "${TMPDIR}/assembly_tracks/${customGenomeAssembly}.gc.bw" ]; then
+    cp ${TMPDIR}/assembly_tracks/${customGenomeAssembly}.gc.bw ${hub_target}/${customGenomeAssembly}/data
+fi
 
 ######################################################################################
 
@@ -217,7 +232,7 @@ echo "hub hub_id_${hub_id}" > "${hub_target}/hub.txt"
 echo "shortLabel ${short_label}" >> "${hub_target}/hub.txt"
 echo "longLabel ${long_label}" >> "${hub_target}/hub.txt"
 echo "genomesFile genomes.txt" >> "${hub_target}/hub.txt"
-echo "email cadley.nyulangone@gmail.com" >> "${hub_target}/hub.txt"
+echo "email cadley.mauranolab@gmail.com" >> "${hub_target}/hub.txt"
 echo "descriptionUrl description.html" >> "${hub_target}/hub.txt"
 
 
@@ -234,8 +249,9 @@ for i in "${genome_array[@]}"; do
     echo " " >> "${hub_target}/genomes.txt"
 done
 
-
-cp ${assemblyBaseDir}/sequences/${customGenomeAssembly}/${customGenomeAssembly}.2bit ${hub_target}/${customGenomeAssembly}/data/${customGenomeAssembly}.2bit
+if [ -f "${assemblyBaseDir}/sequences/${customGenomeAssembly}/${customGenomeAssembly}.2bit" ]; then
+    cp ${assemblyBaseDir}/sequences/${customGenomeAssembly}/${customGenomeAssembly}.2bit ${hub_target}/${customGenomeAssembly}/data/${customGenomeAssembly}.2bit
+fi
 #########################################################
 
 time_stamp=$(date +"%m-%d-%y %T")
@@ -245,8 +261,6 @@ echo "<pre>Hub was constructed at: ${time_stamp} </pre>" >> "${hub_target}/descr
 # Provide a README with the hub link.
 echo
 echo "The UCSC browser url to this hub is:" | tee "${hub_target}/README"
-
-hub_dir=`basename ${hub_target}`
 
 # Construct url for the new hub.
 echo "${URLbase}/${hub_dir}/hub.txt" | tee -a "${hub_target}/README"
